REF: share code ArrayManager/BlockManager (#40338)
